Removed laplace_new, trange, logger
